Add setup command (still broken because depends on wireless command)
